safer code

+++ TopShow/ShowTableViewController.swift
@@ -2,20 +2,14 @@
 import UIKit
 import CoreData
 
-class ShowTablewViewController: UITableViewController, NSFetchedResultsControllerDelegate {
-    
-    
-    // MARK: - IBOutlets
-
-    @IBOutlet weak var sortBySegmentedControl: UISegmentedControl!
-    
-    
+final class ShowTablewViewController: UITableViewController, NSFetchedResultsControllerDelegate {
     
     // MARK: - Variables
     
+    // 1)
     let managedObjectContext = (UIApplication.sharedApplication().delegate as! AppDelegate).managedObjectContext
-    
-    
+
+    //2)
     lazy var fetchedResultsController: NSFetchedResultsController = {
         
         // create fetch request and the sort descriptor
@@ -41,6 +35,9 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
         super.viewDidLoad()
         
         // Perform Fetch
+        // You should really fetch this data asyncronously
+        // but in our app we will likely not have very much data to fetch...
+        // so to simplify things, we will syncronoulsy (that means that things stop till out data is fetched)
         do {
             try fetchedResultsController.performFetch()
         } catch { print("error") }
@@ -65,17 +62,27 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
         Then the changes made in core data are reflected in the UITableView
     */
     func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?) {
-        switch type {
-        case .Delete:
-            self.tableView.deleteRowsAtIndexPaths([indexPath!], withRowAnimation: .Fade)
-        case .Insert:
-            self.tableView.insertRowsAtIndexPaths([newIndexPath!], withRowAnimation: .Fade)
-        case .Update:
-            self.configureCell(self.tableView.cellForRowAtIndexPath(indexPath!)!, indexPath: indexPath!)
-        case .Move:
-            self.tableView.deleteRowsAtIndexPaths([indexPath!], withRowAnimation: .Fade)
-            self.tableView.insertRowsAtIndexPaths([indexPath!], withRowAnimation: .Fade)
+        
+        if type == .Insert, let newindexpath = newIndexPath {
+            self.tableView.insertRowsAtIndexPaths([newindexpath], withRowAnimation: .Fade)
+        } else {
+    
+            if let indexpath = indexPath {
+    
+                switch type {
+                case .Delete:
+                    self.tableView.deleteRowsAtIndexPaths([indexpath], withRowAnimation: .Fade)
+                case .Update:
+                    self.configureCell(self.tableView.cellForRowAtIndexPath(indexPath!)!, indexPath: indexPath!)
+                case .Move:
+                    self.tableView.deleteRowsAtIndexPaths([indexpath], withRowAnimation: .Fade)
+                    self.tableView.insertRowsAtIndexPaths([indexpath], withRowAnimation: .Fade)
+                default:
+                    print("error")
+                }
+            }
         }
+        
     }
     
     
@@ -97,9 +104,12 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
     // MARK: - TabelView Datasource Methods
     
     override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        let sectionInfo = self.fetchedResultsController.sections as [NSFetchedResultsSectionInfo]!
-        let count = sectionInfo[section].numberOfObjects
-        return count
+        if let sectionInfo = self.fetchedResultsController.sections {
+            let count = sectionInfo[section].numberOfObjects
+            return count
+        } else {
+            return 0
+        }
     }
     
     
@@ -112,8 +122,13 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
     
     override func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath) {
         if editingStyle == .Delete {
-            let showToDelete = fetchedResultsController.objectAtIndexPath(indexPath) as! Show
+            guard let showToDelete = fetchedResultsController.objectAtIndexPath(indexPath) as? Show  else {
+                print("error")
+                return
+            }
+    
             managedObjectContext.deleteObject(showToDelete)
+            
             do {
                 try managedObjectContext.save()
             } catch let error as NSError {
@@ -128,16 +143,17 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
     
     func configureCell(cell: UITableViewCell, indexPath: NSIndexPath) {
         let show = self.fetchedResultsController.objectAtIndexPath(indexPath) as! Show
-        if let title = show.title, let rating = show.rating {
+        
+        if let title = show.title {
             cell.textLabel?.text = "\"\(title)\""
-            cell.detailTextLabel?.text = getRating(rating)
+            cell.detailTextLabel?.text = formatRating(show.rating)
             // Gold color for stars
             cell.detailTextLabel?.textColor = UIColor(red: 1.0, green: 0.8431, blue: 0.0, alpha: 1.0)
         }
     }
     
     
-    func getRating(rating: NSNumber) -> String {
+    func formatRating(rating: Int32) -> String {
         switch rating {
         case 5:  return "★★★★★"
         case 4:  return "★★★★"
@@ -155,10 +171,18 @@ class ShowTablewViewController: UITableViewController, NSFetchedResultsControlle
     override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
         
         if segue.identifier == "edit" {
-            if let viewController = (segue.destinationViewController as? UINavigationController)!.topViewController as? Add_EditViewController {
-                if let indexPath = tableView.indexPathForSelectedRow {
-                    let selectedObject = fetchedResultsController.objectAtIndexPath(indexPath) as! Show
-                    viewController.show = selectedObject
+            
+            if let vc = (segue.destinationViewController as? UINavigationController)?.topViewController {
+                
+                if let viewController = vc as? AddEditViewController {
+                
+                    if let indexPath = tableView.indexPathForSelectedRow {
+                    
+                        if let selectedObject = fetchedResultsController.objectAtIndexPath(indexPath) as? Show {
+                            viewController.show = selectedObject
+                        }
+                        
+                    }
                 }
             }
         }
